+++ R/tasmap.R
@@ -1,11 +1,34 @@
-tasmap_layers <- tibble::tribble(
-  ~Name, ~ServiceUrl,
-  "TTSA", "https://services.thelist.tas.gov.au/arcgis/services/Raster/TTSA/MapServer/WMSServer?",
-  "ESgisMapBookPUBLIC", "https://services.thelist.tas.gov.au/arcgis/services/Basemaps/ESgisMapBookPUBLIC/MapServer/WMSServer",
-  "HillshadeGrey", "https://services.thelist.tas.gov.au/arcgis/services/Basemaps/HillshadeGrey/MapServer/WMSServer?",
-  "Tasmap250K", "https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Tasmap250K/MapServer/WMSServer?",
-  "Topographic","https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Topographic/MapServer/WMSServer?"
-)
+# tasmap_layers <- tibble::tribble(
+#   ~Name, ~ServiceUrl,
+#   "TTSA", "https://services.thelist.tas.gov.au/arcgis/services/Raster/TTSA/MapServer/WMSServer?",
+#   "ESgisMapBookPUBLIC", "https://services.thelist.tas.gov.au/arcgis/services/Basemaps/ESgisMapBookPUBLIC/MapServer/WMSServer",
+#   "HillshadeGrey", "https://services.thelist.tas.gov.au/arcgis/services/Basemaps/HillshadeGrey/MapServer/WMSServer?",
+#   "Tasmap250K", "https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Tasmap250K/MapServer/WMSServer?",
+#   "Topographic","https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Topographic/MapServer/WMSServer?"
+# )
+# 
 
+## bases, gsub("/", "_", bases)
+template <- "WMTS:https://services.thelist.tas.gov.au/arcgis/rest/services/%s/MapServer/WMTS/1.0.0/WMTSCapabilities.xml,layer=%s,tilematrixset=default028mm"
+bases <- 
+c("Basemaps/AerialPhoto2020", "Basemaps/AerialPhoto2021", "Basemaps/AerialPhoto2022", 
+  "Basemaps/AerialPhoto2023", "Basemaps/ESgisMapBookPUBLIC", "Basemaps/HillshadeGrey", 
+  "Basemaps/Hillshade", "Basemaps/Orthophoto", "Basemaps/SimpleBasemap", 
+  "Basemaps/Tasmap100K", "Basemaps/Tasmap250K", "Basemaps/Tasmap25K", 
+  "Basemaps/Tasmap500K", "Basemaps/TasmapRaster", "Basemaps/TopographicGrayScale", 
+  "Basemaps/Topographic")
 
-tasmap_ortho <- "WMS:https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Orthophoto/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data%20Boundaries&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454"
+names <- unlist(lapply(strsplit(bases, "/"), "[", 2))
+
+tasmap_sources <- sprintf(template, bases, gsub("/", "_", bases))
+names(tasmap_sources) <- tolower(names)
+## bit more work needed
+tasmap_sources <- c(tasmap_sources, street = "https://services.thelist.tas.gov.au/arcgis/rest/services/Raster/TTSA/MapServer/WMTS/1.0.0/WMTSCapabilities.xml")
+# tasmap_sources <- c(
+#   ortho = "WMS:https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Orthophoto/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data%20Boundaries&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454",
+# street = "WMS:https://services.thelist.tas.gov.au/arcgis/services/Raster/TTSA/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data_Boundaries47811&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454"     ,           
+# mapbook = "WMS:https://services.thelist.tas.gov.au/arcgis/services/Basemaps/ESgisMapBookPUBLIC/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data_Boundaries47811&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454",
+# hillshade = "WMS:https://services.thelist.tas.gov.au/arcgis/services/Basemaps/HillshadeGrey/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data_Boundaries47811&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454"    , 
+# tasmap250 = "WMS:https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Tasmap250K/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data_Boundaries47811&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454"        ,
+# topo = "WMS:https://services.thelist.tas.gov.au/arcgis/services/Basemaps/Topographic/MapServer/WmsServer?SERVICE=WMS&VERSION=1.1.1&REQUEST=GetMap&LAYERS=Data%20Boundaries&SRS=EPSG:4326&BBOX=111.848916,-54.849439,159.145536,-7.800454", 
+# tasmap500 =  "WMTS:https://services.thelist.tas.gov.au/arcgis/rest/services/Basemaps/Tasmap500K/MapServer/WMTS/1.0.0/WMTSCapabilities.xml,layer=Basemaps_Tasmap500K,tilematrixset=default028mm")